requirements.txt
# Core dependencies
web3==6.10.0
python-dotenv==1.0.0
pydantic==2.5.0
eth-account==0.9.0
eth-utils==2.3.0
eth-typing==3.4.0

# Task-specific dependencies
aiohttp==3.8.6
asyncio==3.4.3

# Utilities
loguru==0.7.2
click==8.1.7
tenacity==8.2.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1